Add api service, refactor application service name, remove hello world components.
