+++ frontend/src/components/GrantAccess.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { selectedApplication, selectedApplicationDisplayText } from '../services/ApplicationService'
-import Breadcrumb from './Breadcrumb.vue'
+import { selectedApplication, selectedApplicationDisplayText } from '@/services/ApplicationState'
+import Breadcrumb from '@/components/Breadcrumb.vue'
 
 </script>
