+++ frontend/src/services/ApplicationState.ts
